requirements.txt
python-telegram-bot==21.4
requests>=2.31.0
pillow>=10.0.0
psutil>=5.9.0
rich>=13.7.0
python-dotenv>=1.0.1
